+++ Changelog.md
@@ -1,5 +1,6 @@
 ## 06.06.2024
 1. Added sending alertAlias to OpsGenia
+2. Fix alert handler
 
 ## 05.06.2024
 1. Added Nats